requirements-dev.txt
awsimple
balsa
pressenter2exit
boto3
awscli
github3.py
GitPython
typeguard
pyqt5
sqlitedict
tobool
attrs
appdirs
pref
#
# dev
wheel
black
pytest
coverage
pytest-cov
pyship
flake8
pyqt5-tools
twine
